#579 Compact mode - refactoring
